requirements.txt
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==10.0.0
yolov5==6.2
flask==2.3.3
python-dotenv==1.0.0
exifread==2.4.0
pandas==2.0.3
matplotlib==3.7.1
tensorflow==2.13.0
werkzeug==2.3.7
